feat: show project display name in admin charges filter tag instead of accounting IRI
